+++ backend/src/main/java/edu/fudan/se/sctap_lowcode_tool/model/FusionRuleBranch.java
@@ -14,7 +14,7 @@ public class FusionRuleBranch {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "branch_id")
-    private Long branchId;
+    private int branchId;
 
     @ManyToOne(fetch = FetchType.LAZY)
     @JoinColumn(name = "rule_id", nullable = false)
